Moved functions to a specific socket controller
