+++ server/index.js
@@ -4,6 +4,7 @@ const massive = require('massive');
 const authController = require('./controllers/authController');
 const eventsController = require('./controllers/eventController');
 const peopleController = require('./controllers/peopleController');
+const socketController = require('./controllers/socketController');
 // const session = require('express-session');
 require('dotenv').config();
 // require('dotenv').config();
@@ -14,8 +15,6 @@ const io = require("socket.io")(server);
 
 app.use(bodyParser.json());
 
-// app.use(express.static(`${__dirname}/../build`) );
-
 massive(process.env.CONNECTION_STRING).then(database  => {
     app.set('db', database);
     console.log('database connect is bueno');
@@ -24,46 +23,24 @@ massive(process.env.CONNECTION_STRING).then(database  => {
 // ---------------- SOCKET IO ------------------
 
 io.sockets.on("connection", socket => {
-    console.log("User connected");
+    // user has connected
+    // here we can update them into a current list to view who is online?
 
-    socket.on('join', event => {
-        console.log('joining room?', event.socket_room);
-        socket.join(event.socket_room);
-        app.get('db').get_messages({eventId: event.id})
-        .then( messages => {
-            const messageList = messages.map(m => {
-                return (
-                {
-                    id: m.id,
-                    room: m.socket_room,
-                    name: m.name,
-                    img: m.img,
-                    message: m.message,
-                    time: m.created_time
-                }
-                )
-            })
-            console.log('All Messages', messageList);
-            io.in(event.socket_room).emit('chat-history', messageList);
-        })
-    });
+    // joining a chat room
+    socket.on('join', lobby => socketController.joinRoom(socket, io, lobby, app));
     
-    socket.on('new message', message => {
-        console.log('Message from client', message);
-        app.get('db').create_message({ ...message })
-        .then(event_messages => {
-            const event_message = event_messages[0];
-            console.log(event_message);
-            const m = {
-                id: event_message.id,
-                room: event_message.socket_room,
-                name: event_message.name,
-                img: event_message.img,
-                message: event_message.message,
-                time: event_message.created_time
-            }
-            io.in(message.room).emit('message', m); // return to sender
-        })
+    // sending a message on a chat room
+    socket.on('new message', message => socketController.addMessage(io, message, app));
+
+    // leaving a chat room
+    socket.on('leave', room => {
+        socket.leave(room);
+    })
+
+    // When a user clicks away and the page un-mounts or cuts out this message removes them from all the rooms
+    // Here we can update them in a current list of users online for analytics
+    socket.on('disconnect', () => {
+        socket.leaveAll();
     })
 })
 
@@ -87,6 +64,7 @@ app.delete('/api/events/:id', eventsController.deleteEvent);
 // --------------------  Server Connection  --------------------
 const PORT = 4000
 
+/* Use this for socketIO when we are ready to deploy */
 // const path = require('path')
 // app.get('*', (req, res)=>{
 //
